Optmized art and turned TileMap.LayerDepth to a float
I optimized the artwork by adding sprites to specific packs.

I changed the layerdepth parameter to a float because I thought the
integer steps were too large. Note: rounding errors may occur because of
this and I should probably do something about that.

+++ GelaTina/Assets/Scripts/TileMap.cs
@@ -12,7 +12,7 @@ public class TileMap : MonoBehaviour
     public Vector2 gridSize = new Vector2();
     public int pixelsToUnits = 100;
     public int tileID = 0;
-    public int layerDepth = 0;
+    public float layerDepth = 0;
     public GameObject tiles;
     public bool isSolid;
